code cleaning

+++ Eshop.Controller/src/Controller/OrderController.cs
@@ -2,6 +2,7 @@ using AutoMapper;
 using Eshop.Core.src.Common;
 using Eshop.Service.src.DTO;
 using Eshop.Service.src.ServiceAbstraction;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Eshop.WebApi.Controllers
@@ -23,6 +24,7 @@ namespace Eshop.WebApi.Controllers
 
         // POST: api/Order
         [HttpPost]
+        [Authorize]
         public async Task<ActionResult<OrderReadDTO>> CreateAsync([FromBody] OrderCreateControllerDTO orderDto)
         {
 
@@ -34,7 +36,7 @@ namespace Eshop.WebApi.Controllers
                     throw new InvalidOperationException("Duplicate ProductId detected in order items.");
                 }
             }
-            
+
             var (currentUserId, _) = UserContextHelper.GetUserClaims(HttpContext);
 
             OrderCreateDTO OrderCreateDTO = _mapper.Map<OrderCreateDTO>(orderDto);
@@ -48,61 +50,63 @@ namespace Eshop.WebApi.Controllers
 
         // GET: api/Order
         [HttpGet]
+        [Authorize]
         public async Task<ActionResult<IEnumerable<OrderReadDTO>>> GetAllUserOrders(Guid userId, [FromQuery] QueryOptions options)
         {
+            var (currentUserId, currentUserRole) = UserContextHelper.GetUserClaims(HttpContext);
+
+            if (currentUserId != userId && currentUserRole != "Admin")
+                return Forbid();
+
             var orders = await _orderService.GetAllUserOrdersAsync(userId, options);
             return Ok(_mapper.Map<IEnumerable<OrderReadDTO>>(orders));
         }
 
         // GET: api/Order/{id}
         [HttpGet("{id}")]
+        [Authorize]
         public async Task<ActionResult<OrderReadDTO>> GetOrderById(Guid id)
         {
+
             var order = await _orderService.GetByIdAsync(id);
             if (order == null)
-            {
                 return NotFound();
-            }
+
+            var (currentUserId, currentUserRole) = UserContextHelper.GetUserClaims(HttpContext);
+            if (currentUserId != order.UserId && currentUserRole != "Admin")
+                return Forbid();
+
             return Ok(_mapper.Map<OrderReadDTO>(order));
         }
 
-
         // PUT: api/Order/{id}
         [HttpPut("{id}")]
+        [Authorize]
         public async Task<IActionResult> UpdateOrder(Guid id, OrderUpdateDTO orderUpdateDto)
         {
-            try
-            {
-                var result = await _orderService.UpdateAsync(id, orderUpdateDto);
-                if (!result)
-                {
-                    return NotFound();
-                }
-                return NoContent();
-            }
-            catch (Exception ex)
-            {
-                return BadRequest(new { message = ex.Message });
-            }
+            var (currentUserId, currentUserRole) = UserContextHelper.GetUserClaims(HttpContext);
+            var order = await _orderService.GetByIdAsync(id);
+            if (currentUserId != order.UserId && currentUserRole != "Admin")
+                return Forbid();
+
+            await _orderService.UpdateAsync(id, orderUpdateDto);
+
+            return NoContent();
         }
 
         // DELETE: api/Order/{id}
         [HttpDelete("{id}")]
+        [Authorize]
         public async Task<IActionResult> DeleteOrder(Guid id)
         {
-            try
-            {
-                var result = await _orderService.DeleteByIdAsync(id);
-                if (!result)
-                {
-                    return NotFound();
-                }
-                return NoContent();
-            }
-            catch (Exception ex)
-            {
-                return BadRequest(new { message = ex.Message });
-            }
+            var (currentUserId, currentUserRole) = UserContextHelper.GetUserClaims(HttpContext);
+            var order = await _orderService.GetByIdAsync(id);
+            if (currentUserId != order.UserId && currentUserRole != "Admin")
+                return Forbid();
+
+            await _orderService.DeleteByIdAsync(id);
+
+            return NoContent();
         }
     }
 }
